requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.2
async-timeout==5.0.1
attrs==25.3.0
black==25.9.0
boto3==1.40.44
botocore==1.40.44
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
coverage==7.5.1
cryptography==46.0.3
distro==1.9.0
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==24.3
django-json-widget==2.1.0
django-simple-history==3.5.0
django-storages==1.14.6
django-webpush==0.3.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
Faker==25.0.0
fastapi==0.116.1
flake8==7.0.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
isort==5.13.2
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
mccabe==0.7.0
multidict==6.6.4
mypy_extensions==1.1.0
numpy==1.26.4
openai-agents
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.9
pgvector==0.4.2
py-vapid==1.9.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pywebpush==2.1.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
stripe==14.2.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
twilio==9.7.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
webpush==1.0.6
yarl==1.20.1
google-auth==2.49.1
